Type variables, act III
Unification is hard.

The following example used to type-check:

```
bad : (Fix x. >Int) -> (Fix x. >x)
bad a = a
```

We now properly extend the type context when going under `Fix` binders,
which also means that we have to keep track of the scope of
metavariables and check if solutions have variables that escape their
scope.

For example, the following is not accepted because `x` would escape its
scope:

```
zero : forall a. Fix x. 1 + a
zero = fold (left ())
bad : Fix x. 1 + >x
bad = zero
```